Merge Justin security fixes to shyam security fixes

+++ backend/index.js
@@ -18,6 +18,7 @@ var tfidf = new TfIdf();
 var tokenizer = new natural.WordTokenizer()
 var stopwords = require('stopword');
 const { constrainedMemory } = require('process');
+const { verify } = require('crypto');
 const enStopwords = [
     "i", "me", "my", "myself", "we", "our", "ours", "ourselves", "you", "your", "yours", "yourself", "yourselves",
     "he", "him", "his", "himself", "she", "her", "hers", "herself", "it", "its", "itself", "they", "them", "their",
@@ -155,12 +156,11 @@ async function whoOwnsThis(item, itemValue){
         return await queryTheDatabaseGiveResults("SELECT username FROM ForumPost WHERE id = ?", [itemValue]);
 
     }
+    else if(item === "ForumPostCommentId"){
+        return await queryTheDatabaseGiveResults("SELECT username FROM ForumPostComments WHERE id=?", [itemValue])
+    }
     else if (item === "ForumPostLikeDislikeId"){
         
-    }
-    else if (item === "ForumPostCommentId"){
-        
-
     }
     else if (item === "ForumPostCommentLikeDislikeId"){
         
@@ -272,6 +272,7 @@ app.post("/login", async (req, res)=>{ // secure
         }
     );
 })
+
 app.get('/verify-token', verifyJWTMiddleware, (req, res)=>{ //secure
     return res.json(req.decodedToken);
 })
@@ -633,13 +634,17 @@ app.get("/UserPosts", async (req,res)=>{
 - recalculate tfidf vector for previous posts
 - 
 */
-app.post("/forumPostCreate", async (req, res) => {
+app.post("/forumPostCreate", verifyJWTMiddleware, async (req, res) => {
     const allPosts = req.body.allPosts
     const username = req.body.username
     const title = req.body.title
     const body = req.body.body
 
     const tfidfVector = {}
+
+    if(req.decodedToken.username !== username){
+        return res.status(403).send("Incorrect User")
+    }
 
     //add all previous posts to the corpus
     allPosts.forEach(post => {
@@ -820,7 +825,7 @@ app.get("/forumPostsAll", async (req,res)=>{
 })
 
 //viewing a post by its id
-app.get("/forumPostsById/:postID", async (req,res)=>{
+app.get("/forumPostsById/:postID", verifyJWTMiddleware, async (req,res)=>{
     const id = parseInt(req.params.postID, 10)
     console.log(typeof(id))
     console.log(id)
@@ -843,8 +848,14 @@ app.get("/forumPostsById/:postID", async (req,res)=>{
     })
 })
 
-app.delete("/forumPostDelete/:id", (req,res)=>{
+app.delete("/forumPostDelete/:id", verifyJWTMiddleware, async (req,res)=>{
     const forumPostID = req.params.id
+
+    const postUsername = (await whoOwnsThis("ForumPostId", forumPostID))[0].username;
+    if(req.decodedToken.username !== postUsername){
+        return res.status(403).send("Incorrect User")
+    }
+    
     const query = "DELETE FROM forumpost WHERE id = ?"
     db.query(query, [forumPostID], (err,data)=>{
         if (err) return res.send(err)
@@ -852,8 +863,13 @@ app.delete("/forumPostDelete/:id", (req,res)=>{
     });
 });
 
-app.put("/editForumPost/:id", (req, res) =>{
+app.put("/editForumPost/:id", verifyJWTMiddleware, async (req, res) =>{
     const forumPostID = req.params.id
+    const postUsername = (await whoOwnsThis("ForumPostId", forumPostID))[0].username;
+    if(req.decodedToken.username !== postUsername){
+        return res.status(403).send("Incorrect User")
+    }
+
     const query = "UPDATE forumpost SET body=? WHERE id=?"
     db.query(query, [req.body.newBody, forumPostID], (err, data)=>{
         if(err){
@@ -974,12 +990,15 @@ app.delete("/forumPostDeleteLikeDislike/", async (req,res)=>{
     });
 });
 
-app.post("/forumPostComment/:id", (req, res) => {
+app.post("/forumPostComment/:id", verifyJWTMiddleware, (req, res) => {
     const forumPostID = parseInt(req.params.id, 10)
     //debugging purposes
     const username = req.body.username
     const body = req.body.body
 
+    if(req.decodedToken.username !== username){
+        return res.status(403).send("Incorrect User")
+    }
 
    db.query("INSERT INTO forumpostcomments(forum_post_id, username, body, comment_timestamp) VALUES (?, ?, ?, NOW())", [forumPostID, username, body], function (err){
     if(err){
@@ -990,7 +1009,7 @@ app.post("/forumPostComment/:id", (req, res) => {
     })
 })
 
-app.get("/getForumPostCommentByID/:id", (req, res) =>{
+app.get("/getForumPostCommentByID/:id", verifyJWTMiddleware, (req, res) =>{
     const forumPostCommentID = parseInt(req.params.id, 10)
     db.query("SELECT * FROM FORUMPOSTCOMMENTS WHERE id=?", [forumPostCommentID], function(err, data){
         if(err){
@@ -1002,8 +1021,14 @@ app.get("/getForumPostCommentByID/:id", (req, res) =>{
     })
 })
 
-app.put("/deleteForumPostComment/:commentID", (req, res) =>{
+
+app.put("/deleteForumPostComment/:commentID",verifyJWTMiddleware, async (req, res) =>{
     const commentID = parseInt(req.params.commentID, 10)
+    const commentUser = (await whoOwnsThis("ForumPostCommentId", commentID))[0].username;
+    if(req.decodedToken.username !== commentUser){
+        return res.status(403).send("Incorrect User")
+    }
+
     console.log(commentID)
     db.query("UPDATE FORUMPOSTCOMMENTS SET DELETED=? WHERE id=?", [true, commentID], function(err){
         if(err){
@@ -1015,11 +1040,16 @@ app.put("/deleteForumPostComment/:commentID", (req, res) =>{
     })
 })
 
-app.put("/editForumPostComment/:commentID", (req, res) =>{
+app.put("/editForumPostComment/:commentID", verifyJWTMiddleware, async (req, res) =>{
     console.log("HELLO")
     const commentID = parseInt(req.params.commentID, 10)
     const editedBody = req.body.editedBody
+    const commentUser = (await whoOwnsThis("ForumPostCommentId", commentID))[0].username;
     console.log(editedBody)
+    if(req.decodedToken.username !== commentUser){
+        return res.status(403).send("Incorrect User")
+    }
+
     db.query("UPDATE FORUMPOSTCOMMENTS SET body=? WHERE id=?", [editedBody, commentID], function(err){
         if(err){
             console.log(err)
@@ -1030,7 +1060,7 @@ app.put("/editForumPostComment/:commentID", (req, res) =>{
     })
 })
 
-app.get("/getForumPostComments", async(req, res) =>{
+app.get("/getForumPostComments", verifyJWTMiddleware, async(req, res) =>{
     const username = req.query.username
     db.query("SELECT * FROM ForumPostComments WHERE username=?", [username], function(err, data){
         if(err){
@@ -1161,7 +1191,7 @@ app.get("/forumPostCommentsDislikedByUser/", async (req,res)=>{
     });
 });
 
-app.get("/forumPostParentCommentGetByID/:id", (req, res) =>{
+app.get("/forumPostParentCommentGetByID/:id", verifyJWTMiddleware, (req, res) =>{
     const forumPostID = parseInt(req.params.id, 10)
     db.query("SELECT * FROM forumpostcomments WHERE forum_post_id=? AND parent_comment_id IS NULL", [forumPostID], function (err, data){
     if(err){
@@ -1171,12 +1201,16 @@ app.get("/forumPostParentCommentGetByID/:id", (req, res) =>{
     })
 })
 
-app.post("/forumPostCommentReply/:id/:commentID", (req, res) => {
+app.post("/forumPostCommentReply/:id/:commentID", verifyJWTMiddleware, (req, res) => {
     const forumPostID = parseInt(req.params.id, 10)
     const commentID = parseInt(req.params.commentID, 10)
     const username = req.body.username
     const body = req.body.body
     const timestamp = new Date().toISOString()
+
+    if(req.decodedToken.username != username){
+        return res.status(403).send("Incorrect User")
+    }
 
     console.log("FORUM POST COMMENT REPLY FOR POST " + forumPostID)
     console.log("TYPE OF FORUM POST ID " + typeof(forumPostID))
@@ -1252,7 +1286,7 @@ app.post("/forumPostCommentReply/:id/:commentID", (req, res) => {
     })
 })
 
-app.get("/forumPostParentGetReplies/:id/:commentID", (req, res) =>{
+app.get("/forumPostParentGetReplies/:id/:commentID", verifyJWTMiddleware, (req, res) =>{
     const forumPostID = parseInt(req.params.id, 10)
     const parentCommentID = parseInt(req.params.commentID, 10)
 
@@ -1270,7 +1304,7 @@ app.get("/forumPostParentGetReplies/:id/:commentID", (req, res) =>{
     })
 })
 
-app.get("/forumPostSearch/:searchTitle", (req, res) => {
+app.get("/forumPostSearch/:searchTitle", verifyJWTMiddleware, (req, res) => {
     const searchTitle = req.params.searchTitle
     db.query("SELECT * FROM forumpost WHERE title LIKE ?",['%' + searchTitle + '%'], function (err, data){
         if(err){
@@ -1309,7 +1343,7 @@ function cosineSimilarity(tfidfVector1, tfidfVector2) {
 to see if a post should be recommended, take the consine similarity of the tfidf_vector
 if it meets the threshold, recommend it
 */
-app.get("/forumPostRecommendedPost/:postID", (req, res) =>{
+app.get("/forumPostRecommendedPost/:postID", verifyJWTMiddleware, (req, res) =>{
     const postID = req.params.postID
     const recommendedPosts = []
     db.query("SELECT * from forumpost WHERE ID=?", [postID], function(err, data){
